Multi-domain alignment strategies.

+++ Libs/Mesh/MeshUtils.h
@@ -32,7 +32,7 @@ public:
   static PhysicalRegion boundingBox(const std::vector<std::reference_wrapper<const Mesh>>& meshes, bool center = false);
 
   /// determine the reference mesh
-  static int findReferenceMesh(std::vector<Mesh> &meshes);
+  static size_t findReferenceMesh(std::vector<Mesh> &meshes);
 
   /// generates and adds normals for points and faces for each mesh in given set of meshes
   static void generateNormals(const std::vector<std::reference_wrapper<Mesh>>& meshes, bool forceRegen = false);
